requirements.txt
# Core dependencies
openai>=1.0.0
jsonlines>=3.0.0
python-dotenv>=1.0.0

# Retrieval & Embedding
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0
torch>=2.0.0

# BM25 (optional, for hybrid retrieval)
rank_bm25>=0.2.2

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Data quality
tabulate>=0.9.0

